IntegrationTestExtension: Move client-independent helpers to a dedicated class
